+++ qiime/support_files/R/loaddata.r
@@ -12,7 +12,7 @@
     close(f)
     # read the rest of the table
     datatable <- read.table(filepath,sep='\t',skip=header.index, comment='#',quote='"',
-                        head=F,row.names=1,check=FALSE)
+                        head=F,row.names=1,check=FALSE,strip.white=TRUE)
     
     # set column names using header
     colnames(datatable) <- header[-1]