+++ ChangeLog.md
@@ -7,6 +7,7 @@ QIIME 1.7.0-dev (changes since QIIME 1.7.0 go here)
 * transform_coordinate_matrices.py can now take more than two input coordinate matrices. When used this way, the first coordinate matrix will be treated as the reference, and the 2nd through nth will be compared against that reference. The output file names, which were all previously hard-coded, are now generated on the fly for clarity of the results.
 * split_libraries_fastq.py can now handle per-sample, non-barcoded fastq files. Some sequencing centers are now providing data in this way - if this becomes more common, we'll want to make this more convenient, but for now it's possible.
 * Added a parallel merge OTUs method that will combine OTU tables in parallel where possible.
+* Added identify_paired_differences.py to support paired difference (i.e., Pre/Post) testing as discussed in issue #1040.
 * Required biom-format version is now 1.2.0.
 
 QIIME 1.7.0 (14 May 2013)